tested likes and comment models. added service folder

+++ backend/src/controller/CommentController.ts
@@ -5,9 +5,13 @@ import {
   commentInsertZodSchema,
   commentSelectSchema,
 } from "../types/comment";
-import CommentsModel from "../models/CommentsModel";
+import CommentsModel, { CommentModel } from "../models/CommentsModel";
 import { Constants } from "../Constants";
 import z from "zod";
+import { getUserDataById } from "../service/userService";
+import { getReciepById } from "../service/recipeService";
+import { Res } from "../types/res";
+import { getCommentByCommentId, getCommentsByRecipeId } from "../service/commentService";
 
 export const addComment = async (
   req: Request<unknown, unknown, CommentInsert>,
@@ -15,6 +19,20 @@ export const addComment = async (
 ) => {
   try {
     const validComment = commentInsertZodSchema.parse(req.body);
+    validComment.createdAt = new Date();
+    const [user, recipe] = await Promise.all([
+      getUserDataById(validComment.authorId),
+      getReciepById(validComment.recipeId),
+    ]);
+
+    if (!user || !recipe) {
+      return res.status(Constants.NOT_FOUND).json({
+        isSuccess: false,
+        data: {
+          message: "Either recipe or user not found",
+        },
+      });
+    }
     const savedComment = await CommentsModel.create(validComment);
     if (savedComment) {
       return res.status(Constants.CREATED).json({
@@ -32,6 +50,14 @@ export const addComment = async (
       },
     });
   } catch (error) {
+    if(error.code.toString() === "11000"){
+      return res.status(Constants.VALIDATION_ERROR).json({
+        isSuccess: false,
+        data: {
+          message: "duplicate comment",
+        },
+      })
+    }
     return res.status(Constants.SERVER_ERROR).json({
       isSuccess: false,
       data: {
@@ -48,7 +74,7 @@ export const editComment = async (
   try {
     const validComment = commentSelectSchema.parse(req.body);
     const id = req.params.id;
-    const savedComment = await CommentsModel.findById(id);
+    const savedComment = await getCommentByCommentId(id);
     if (!savedComment) {
       return res.status(Constants.NOT_FOUND).json({
         isSuccess: false,
@@ -71,16 +97,25 @@ export const editComment = async (
 
 export const deleteComment = async (
   req: Request<{ id: string }>,
-  res: Response
+  res: Response<Res<any>>
 ) => {
   try {
     const validId = z.string().parse(req.params.id);
+    const comment = await getCommentByCommentId(validId);
+    if (!comment) {
+      return res.status(Constants.NOT_FOUND).json({
+        isSuccess: false,
+        data: {
+          message: "Comment not found",
+        },
+      });
+    }
     const deletedComment = await CommentsModel.findByIdAndDelete(validId);
     if (!deleteComment) {
       return res.status(Constants.VALIDATION_ERROR).json({
         isSuccess: false,
         data: {
-          message: "Comment was not deleted",
+          message: "Comment not found",
         },
       });
     }
@@ -88,7 +123,7 @@ export const deleteComment = async (
       isSuccess: true,
       data: {
         message: "Comment deleted successfully",
-        data: deletedComment,
+        result: deletedComment,
       },
     });
   } catch (error) {
@@ -99,4 +134,37 @@ export const deleteComment = async (
       },
     });
   }
+};
+
+export const commentsByRecipeId = async (
+  req: Request<{ id: string }, unknown>,
+  res: Response<Res<CommentModel[]>>
+) => {
+  try {
+    const recipe = await getReciepById(req.params.id);
+    if (!recipe) {
+      return res.status(Constants.NOT_FOUND).json({
+        isSuccess: false,
+        data: {
+          message: "Recipe not found",
+        },
+      });
+    }
+    const comments = await getCommentsByRecipeId(req.params.id);
+    if (!comments) {
+      return res.status(Constants.NOT_FOUND).json({
+        isSuccess: false,
+        data: {
+          message: "Comments not found",
+        },
+      });
+    }
+    return res.status(Constants.OK).json({
+      isSuccess: true,
+      data: {
+        message: "Comments found",
+        result: comments,
+      },
+    });
+  } catch (error) {}
 };
